Move ProfileFields to its own file in entity for reuse
